fix on interaction between codegen and process
